A lot of restructuring. Still far from perfect / correct, but a lot better than before
