--- .vscode/settings.json
@@ -1,14 +1,0 @@
-{
-    "files.associations": {
-        "*.tcc": "cpp",
-        "algorithm": "cpp",
-        "memory": "cpp",
-        "random": "cpp",
-        "fstream": "cpp",
-        "iosfwd": "cpp",
-        "istream": "cpp",
-        "ostream": "cpp",
-        "sstream": "cpp",
-        "streambuf": "cpp"
-    }
-}
