+++ lib/images/images.h
@@ -1,0 +1,35 @@
+#pragma once
+
+#include <Arduino.h>
+
+// BITMAP-BILDER BITMAP-BILDER BITMAP-BILDER BITMAP-BILDER BITMAP-BILDER
+
+extern const uint8_t settingsbutton10[] PROGMEM;
+extern const uint8_t settingsbutton40[] PROGMEM;
+extern const uint8_t settingsbutton30[] PROGMEM;
+
+extern const uint8_t espresso_cup40[] PROGMEM;
+
+extern const uint8_t espresso_cup30[] PROGMEM;
+
+extern const uint8_t back_10[] PROGMEM;
+
+extern const uint8_t round_10[] PROGMEM;
+extern const uint8_t roundfilled_10[]PROGMEM;
+
+extern const uint8_t home_10[] PROGMEM;
+
+extern const uint8_t home_10_filled[] PROGMEM;
+
+extern const uint8_t espresso_10[] PROGMEM;
+extern const uint8_t trash_10 [] PROGMEM;
+
+extern const uint8_t trash_10_filled [] PROGMEM;
+
+extern const uint8_t trash_20 [] PROGMEM;
+
+extern const uint8_t home_20 [] PROGMEM;
+
+extern const uint8_t hourglass_10 [] PROGMEM;
+
+extern const uint8_t hourglass_10_filled [] PROGMEM;
