+++ lib/ui/ui.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <Arduino.h>
+
+void displayTemperature();
+void displaySettings(int index);
+void displayMainMenu(byte index);
+void displaypowerStates(byte index);
