+++ platformio.ini
@@ -8,6 +8,23 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+; [env:esp32doit-devkit-v1]
+; framework = arduino
+; platform = espressif32
+; platform_packages =
+;   ; use a special branch
+; ;   framework-arduinoespressif32 @ https://github.com/espressif/arduino-esp32.git#idf-release/v3.3
+
+; board = esp32doit-devkit-v1
+; ;monitor_speed=115200
+; lib_deps = 
+; 	br3ttb/PID @ ^1.2.1
+; 	milesburton/DallasTemperature @ ^3.9.1
+; 	knolleary/PubSubClient@^2.8
+; 	thingpulse/ESP8266 and ESP32 OLED driver for SSD1306 displays @ ^4.2.0
+
+
+
 [env:esp32doit-devkit-v1]
 platform = espressif32
 platform_packages =
@@ -15,38 +32,11 @@ platform_packages =
 board = esp32doit-devkit-v1
 framework = arduino
 ;monitor_speed=115200
+lib_ldf_mode = deep+
 lib_deps = 
 	br3ttb/PID @ ^1.2.1
-	squix78/ESP8266_SSD1306 @ ^4.1.0
 	milesburton/DallasTemperature @ ^3.9.1
 	knolleary/PubSubClient@^2.8
-
-; [env:esp32doit-devkit-v1]
-; ; [env:esp32dev]
-; ; [env:latest_stable]
-; ;[env:lolin32]
-; ; [env:upstream_develop]
-; framework = espidf, arduino 
-; ; platform = https://github.com/platformio/platform-espressif32.git
-; ; platform_packages =
-; ; 	esp-idf @ https://github.com/espressif/esp-idf.git
-; ; 	; framework-espidf @ https://github.com/espressif/esp-idf.git
-; platform = espressif32
-; platform_packages =
-;   	; use a special branch
-;   	; framework-espidf @ https://github.com/espressif/arduino-esp32.git#idf-release/v4.0
-;     ; platformio/framework-arduinoespressif32 @ https://github.com/espressif/arduino-esp32
-; 	; platformio/framework-arduinoespressif32 @ https://github.com/espressif/arduino-esp32#idf-release/v4.0
-; 	framework-arduinoespressif32 @ https://github.com/marcovannoord/arduino-esp32.git#idf-release/v4.0
-; 	esp-idf @ https://github.com/espressif/esp-idf
-; ; board = lolin32
-; board = esp32doit-devkit-v1
-; ; board = esp32dev
-; ;monitor_speed=115200
-; lib_deps = 
-; 	br3ttb/PID @ ^1.2.1
-; 	; squix78/ESP8266_SSD1306 @ ^4.1.0
-; 	milesburton/DallasTemperature @ ^3.9.1
-; 	knolleary/PubSubClient@^2.8
-; 	thingpulse/ESP8266 and ESP32 OLED driver for SSD1306 displays @ ^4.2.0
+	thingpulse/ESP8266 and ESP32 OLED driver for SSD1306 displays @ ^4.2.0
+	; arduino-libraries/WiFi @ ^1.2.7
