+++ raw_bitmaps/circle_10x10.xbm
