+++ raw_bitmaps/circle_10x10_filled.xbm
