+++ raw_bitmaps/espresso_cup2_40x40.xbm
