+++ raw_bitmaps/espresso_cup_10x10.xbm
