+++ raw_bitmaps/espresso_cup_10x10_filled.xbm
