+++ raw_bitmaps/espresso_cup_30x30.xbm
