+++ raw_bitmaps/espresso_cup_40x40.xbm
