+++ raw_bitmaps/home_10x10.xbm
