+++ raw_bitmaps/hourclass_10x10.xbm
