+++ raw_bitmaps/hourclass_10x10_filled.xbm
