+++ raw_bitmaps/settings_40x40.xbm
