+++ raw_bitmaps/trash_10x10.xbm
