+++ raw_bitmaps/trash_10x10_filled.xbm
