+++ raw_bitmaps/trash_20x20.xbm
